refactored TOE GetHelp component into a shared comp and swapped in TOE and MEB.
