+++ src/applications/my-education-benefits-apps/my-education-benefits/components/FormFooter.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import GetFormHelp from './GetFormHelp';
+import GetHelp from '../../shared/components/GetHelp';
 
 export default function FormFooter() {
   return (
     <div className="help-footer-box">
       <h2 className="help-heading">Need help?</h2>
-      <GetFormHelp />
+      <GetHelp />
     </div>
   );
 }
