--- src/applications/my-education-benefits-apps/my-education-benefits/components/GetFormHelp.jsx
@@ -1,22 +1,0 @@
-import React from 'react';
-
-function GetFormHelp() {
-  return (
-    <>
-      <p className="help-talk">
-        If you need help with your application or have questions about
-        enrollment or eligibility, call our Education Call Center at{' '}
-        <a href="tel:888-442-4551">888-442-4551 (711)</a>. We're here Monday
-        through Friday, 8:00 a.m. to 7:00 p.m. ET. If you're outside the U.S.,
-        call us at <a href="tel:001-918-781-5678">001-918-781-5678</a>.
-      </p>
-      <p>
-        If you have technical difficulties using this online application, call
-        our MyVA411 main information line at{' '}
-        <a href="tel:800-698-2411">800-698-2411 (711)</a>. We're here 24/7.
-      </p>
-    </>
-  );
-}
-
-export default GetFormHelp;
